move block index to its own file
