+++ zebra-state/Cargo.toml
@@ -14,9 +14,14 @@ eyre = "0.4.2"
 futures = "0.3.5"
 lazy_static = "1.4.0"
 hex = "0.4.2"
+tracing-error = "0.1.2"
+tracing-subscriber = "0.2.5"
 
 [dev-dependencies]
 color-eyre = "0.3.4"
 eyre = "0.4.2"
 tokio = { version = "0.2.21", features = ["full"] }
 zebra-test-vectors = { path = "../zebra-test-vectors/" }
+spandoc = { git = "https://github.com/yaahc/spandoc.git" }
+tracing = "0.1.15"
+tracing-futures = "0.2.4"
