+++ zebra-state/src/in_memory.rs
@@ -1,56 +1,18 @@
 use super::{Request, Response};
 use futures::prelude::*;
 use std::{
-    collections::{BTreeMap, HashMap},
     error::Error,
     future::Future,
     pin::Pin,
-    sync::Arc,
     task::{Context, Poll},
 };
 use tower::{buffer::Buffer, Service};
-use zebra_chain::{
-    block::{Block, BlockHeaderHash},
-    types::BlockHeight,
-};
+
+mod block_index;
 
 #[derive(Default)]
 struct ZebraState {
-    by_hash: HashMap<BlockHeaderHash, Arc<Block>>,
-    by_height: BTreeMap<BlockHeight, Arc<Block>>,
-}
-
-impl ZebraState {
-    fn insert(&mut self, block: impl Into<Arc<Block>>) {
-        let block = block.into();
-        let hash = block.as_ref().into();
-        let height = block.coinbase_height().unwrap();
-
-        assert!(
-            self.by_hash.insert(hash, block.clone()).is_none(),
-            "blocks shouldn't have the same hash"
-        );
-        assert!(
-            self.by_height.insert(height, block).is_none(),
-            "blocks with the same height are currently unsupported"
-        );
-    }
-
-    fn get(&mut self, query: impl Into<BlockQuery>) -> Option<Arc<Block>> {
-        match query.into() {
-            BlockQuery::ByHash(hash) => self.by_hash.get(&hash),
-            BlockQuery::ByHeight(height) => self.by_height.get(&height),
-        }
-        .cloned()
-    }
-
-    fn tip(&self) -> Option<Arc<Block>> {
-        self.by_height
-            .iter()
-            .next_back()
-            .map(|(_key, value)| value)
-            .cloned()
-    }
+    index: block_index::BlockIndex,
 }
 
 impl Service<Request> for ZebraState {
@@ -66,11 +28,12 @@ impl Service<Request> for ZebraState {
     fn call(&mut self, req: Request) -> Self::Future {
         match req {
             Request::AddBlock { block } => {
-                self.insert(block);
+                self.index.insert(block);
                 async { Ok(Response::Added) }.boxed()
             }
             Request::GetBlock { hash } => {
                 let result = self
+                    .index
                     .get(hash)
                     .map(|block| Response::Block { block })
                     .ok_or_else(|| "block could not be found".into());
@@ -79,6 +42,7 @@ impl Service<Request> for ZebraState {
             }
             Request::GetTip => {
                 let result = self
+                    .index
                     .tip()
                     .map(|block| block.as_ref().into())
                     .map(|hash| Response::Tip { hash })
@@ -87,23 +51,6 @@ impl Service<Request> for ZebraState {
                 async move { result }.boxed()
             }
         }
-    }
-}
-
-enum BlockQuery {
-    ByHash(BlockHeaderHash),
-    ByHeight(BlockHeight),
-}
-
-impl From<BlockHeaderHash> for BlockQuery {
-    fn from(hash: BlockHeaderHash) -> Self {
-        Self::ByHash(hash)
-    }
-}
-
-impl From<BlockHeight> for BlockQuery {
-    fn from(height: BlockHeight) -> Self {
-        Self::ByHeight(height)
     }
 }
